chore: improve static analysis setup local+CI
As it was hard to execute the static analysis locally while also keeping
it working in CI, we decided to address that later, which happens in
this commit.

We use '--file' and $PWD to generate an absolute path based on the
working directory (the workspace!), which works properly locally and in
CI.

We need to remove the '-fanalyzer' flag in the lib compilation though,
as clangsa fails with an unknown flag error and does not analyze the
files.

This resolves #14

Signed-off-by: Stefan Kraus <stefan.kraus@ul.com>