chore: exclude crypt 'update/finish' from scope
As we right now do not use these function in our sample as our messages
for demonstration are quite small, and we prefer having proper key
handling first, we remove these to reduce the scope of testing and
re-testing on changes, and re-add these functions with test later.

The task is documented in #18 so we don't forget it.

Signed-off-by: Stefan Kraus <stefan.kraus@ul.com>