feat: change _Init functions to match spec
Previosly, Encrypt could be called without EncryptInit, and
also be called multiple times without an EncryptInit between.

This commit changes this to match the specification that enforces an
EncryptInit before each Encrypt calls.

Same applies for DecryptInit and Decrypt.
Corresponding integration tests are added in the next commit.

This is the implementation corresponding to #37.

Signed-off-by: Stefan Kraus <stefan.kraus@ul.com>